flutter app included

+++ Backend/models/Agri.js
@@ -3,16 +3,16 @@ import { sequelize } from '../config/database.js';
 
 const SoilReport = sequelize.define('SoilReport', {
     id: { type: DataTypes.UUID, defaultValue: DataTypes.UUIDV4, primaryKey: true },
-    ph: { type: DataTypes.DECIMAL },
-    nitrogen: { type: DataTypes.DECIMAL },
-    phosphorus: { type: DataTypes.DECIMAL },
-    potassium: { type: DataTypes.DECIMAL },
-    organicMatter: { type: DataTypes.DECIMAL },
+    ph: { type: DataTypes.FLOAT },
+    nitrogen: { type: DataTypes.FLOAT },
+    phosphorus: { type: DataTypes.FLOAT },
+    potassium: { type: DataTypes.FLOAT },
+    organicMatter: { type: DataTypes.FLOAT },
     micronutrients: { type: DataTypes.JSONB, defaultValue: {} },
     recommendedFertilizer: { type: DataTypes.STRING },
     suitableCrops: { type: DataTypes.ARRAY(DataTypes.STRING), defaultValue: [] },
     soilStatus: { type: DataTypes.STRING },
-    suitabilityPct: { type: DataTypes.DECIMAL },
+    suitabilityPct: { type: DataTypes.FLOAT },
     reportFile: { type: DataTypes.STRING },
     status: { type: DataTypes.STRING, defaultValue: 'Pending' },
     nextTestDate: { type: DataTypes.DATE }
@@ -37,10 +37,10 @@ const MarketPrice = sequelize.define('MarketPrice', {
     cropName: { type: DataTypes.STRING, allowNull: false },
     district: { type: DataTypes.STRING, allowNull: false },
     mandi: { type: DataTypes.STRING, defaultValue: 'Local Mandi' },
-    minPrice: { type: DataTypes.DECIMAL },
-    maxPrice: { type: DataTypes.DECIMAL },
-    modalPrice: { type: DataTypes.DECIMAL, allowNull: false },
-    change: { type: DataTypes.DECIMAL, defaultValue: 0 },
+    minPrice: { type: DataTypes.FLOAT },
+    maxPrice: { type: DataTypes.FLOAT },
+    modalPrice: { type: DataTypes.FLOAT, allowNull: false },
+    change: { type: DataTypes.FLOAT, defaultValue: 0 },
     arrivalDate: { type: DataTypes.DATE, defaultValue: DataTypes.NOW },
     variety: { type: DataTypes.STRING, defaultValue: 'Standard' },
     source: { type: DataTypes.STRING, defaultValue: 'agmarknet' }
@@ -54,7 +54,7 @@ const MarketHistory = sequelize.define('MarketHistory', {
     id: { type: DataTypes.UUID, defaultValue: DataTypes.UUIDV4, primaryKey: true },
     crop: { type: DataTypes.STRING, allowNull: false },
     district: { type: DataTypes.STRING, allowNull: false },
-    price: { type: DataTypes.DECIMAL, allowNull: false },
+    price: { type: DataTypes.FLOAT, allowNull: false },
     date: { type: DataTypes.DATE, defaultValue: DataTypes.NOW }
 }, {
     tableName: 'market_history',
@@ -65,7 +65,7 @@ const MarketHistory = sequelize.define('MarketHistory', {
 const PriceAlert = sequelize.define('PriceAlert', {
     id: { type: DataTypes.UUID, defaultValue: DataTypes.UUIDV4, primaryKey: true },
     crop: { type: DataTypes.STRING, allowNull: false },
-    targetPrice: { type: DataTypes.DECIMAL, allowNull: false },
+    targetPrice: { type: DataTypes.FLOAT, allowNull: false },
     condition: { type: DataTypes.STRING, allowNull: false },
     status: { type: DataTypes.STRING, defaultValue: 'Active' },
     lastNotified: { type: DataTypes.DATE }
